tested with 2 layers

+++ models/pytorch/lstm_predict.py
@@ -68,7 +68,7 @@ def predict_fn(input_data, model):
 
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
-    data = torch.from_numpy(input_data.astype('float32'))
+    data = torch.from_numpy(input_data)
     data = data.to(device)
 
     h = model.init_hidden(1024)
